requirements.txt

# OpenAI's LangChain integration
langchain-openai

# Other dependencies
streamlit
python-docx
docx2txt
gradio
cloudinary
requests
beautifulsoup4
langchain-community
langchain
pypdf
faiss-cpu
python-dotenv